Bot/requirements.txt
#pip install -r requirements.txt для скачки с помощью PowerShell
aiogram==3.16.0
pandas==2.3.3
openpyxl==3.1.5
xlrd==2.0.2
python-dotenv==1.0.1
